+++ ntex/src/web/types/json.rs
@@ -7,7 +7,7 @@ use std::task::{Context, Poll};
 use std::{fmt, ops};
 
 use bytes::BytesMut;
-use futures::future::{err, ok, FutureExt, LocalBoxFuture, Ready};
+use futures::future::{ready, FutureExt, LocalBoxFuture, Ready};
 use futures::StreamExt;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
@@ -16,7 +16,7 @@ use serde::Serialize;
 use crate::http::encoding::Decoder;
 use crate::http::header::CONTENT_LENGTH;
 use crate::http::{HttpMessage, Payload, Response, StatusCode};
-use crate::web::error::{ErrorRenderer, JsonError, JsonPayloadError};
+use crate::web::error::{ErrorRenderer, JsonError, JsonPayloadError, WebResponseError};
 use crate::web::{FromRequest, HttpRequest, Responder};
 
 /// Json helper
@@ -117,19 +117,24 @@ where
     }
 }
 
-impl<T: Serialize, Err: ErrorRenderer> Responder<Err> for Json<T> {
+impl<T: Serialize, Err: ErrorRenderer> Responder<Err> for Json<T>
+where
+    Err::Container: From<JsonError>,
+{
     type Error = JsonError;
-    type Future = Ready<Result<Response, Self::Error>>;
+    type Future = Ready<Response>;
 
-    fn respond_to(self, _: &HttpRequest) -> Self::Future {
+    fn respond_to(self, req: &HttpRequest) -> Self::Future {
         let body = match serde_json::to_string(&self.0) {
             Ok(body) => body,
-            Err(e) => return err(e),
+            Err(e) => return ready(e.error_response(req)),
         };
 
-        ok(Response::build(StatusCode::OK)
-            .content_type("application/json")
-            .body(body))
+        ready(
+            Response::build(StatusCode::OK)
+                .content_type("application/json")
+                .body(body),
+        )
     }
 }
 
@@ -414,7 +419,7 @@ mod tests {
         let j = Json(MyObject {
             name: "test".to_string(),
         });
-        let resp = respond_to(j, &req).await.unwrap();
+        let resp = respond_to(j, &req).await;
         assert_eq!(resp.status(), StatusCode::OK);
         assert_eq!(
             resp.headers().get(header::CONTENT_TYPE).unwrap(),